requirements.txt
streamlit>=1.35.0
streamlit-quill>=0.0.1
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
langchain>=0.3.0
langchain-groq>=0.2.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
